changed total media thing. added label under.
seems to look better when having it under having it be a "label" instead of a title
should probably be that the cards have a title and that the data inside has a label.

will need to change layout completely inside the categories to make it work, but it probably add some extra heiarchy and look way better.