Moved it all to svelte, took me 7 hours ish.
todo:
fix so the rendering of the stats is done with svelte, instead of a big js file.
change charts to be made by me using html and magic.
or if there is a charts plugin for svelte.

handle the data send to server to displayt the statistics as in point 1.